+++ bailian-20231229/src/Models/BatchUpdateFileTagRequest/fileInfos.php
@@ -1,0 +1,79 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Bailian\V20231229\Models\BatchUpdateFileTagRequest;
+
+use AlibabaCloud\Dara\Model;
+
+class fileInfos extends Model
+{
+    /**
+     * @var string
+     */
+    public $fileId;
+
+    /**
+     * @var string[]
+     */
+    public $tags;
+    protected $_name = [
+        'fileId' => 'FileId',
+        'tags' => 'tags',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->tags)) {
+            Model::validateArray($this->tags);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->fileId) {
+            $res['FileId'] = $this->fileId;
+        }
+
+        if (null !== $this->tags) {
+            if (\is_array($this->tags)) {
+                $res['tags'] = [];
+                $n1 = 0;
+                foreach ($this->tags as $item1) {
+                    $res['tags'][$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['FileId'])) {
+            $model->fileId = $map['FileId'];
+        }
+
+        if (isset($map['tags'])) {
+            if (!empty($map['tags'])) {
+                $model->tags = [];
+                $n1 = 0;
+                foreach ($map['tags'] as $item1) {
+                    $model->tags[$n1] = $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        return $model;
+    }
+}