+++ bailian-20231229/src/Models/BatchUpdateFileTagResponseBody/data/updateFileTagResultList.php
@@ -1,0 +1,90 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Bailian\V20231229\Models\BatchUpdateFileTagResponseBody\data;
+
+use AlibabaCloud\Dara\Model;
+
+class updateFileTagResultList extends Model
+{
+    /**
+     * @var string
+     */
+    public $errorCode;
+
+    /**
+     * @var string
+     */
+    public $errorMessage;
+
+    /**
+     * @var string
+     */
+    public $fileId;
+
+    /**
+     * @var bool
+     */
+    public $success;
+    protected $_name = [
+        'errorCode' => 'ErrorCode',
+        'errorMessage' => 'ErrorMessage',
+        'fileId' => 'FileId',
+        'success' => 'Success',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->errorCode) {
+            $res['ErrorCode'] = $this->errorCode;
+        }
+
+        if (null !== $this->errorMessage) {
+            $res['ErrorMessage'] = $this->errorMessage;
+        }
+
+        if (null !== $this->fileId) {
+            $res['FileId'] = $this->fileId;
+        }
+
+        if (null !== $this->success) {
+            $res['Success'] = $this->success;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['ErrorCode'])) {
+            $model->errorCode = $map['ErrorCode'];
+        }
+
+        if (isset($map['ErrorMessage'])) {
+            $model->errorMessage = $map['ErrorMessage'];
+        }
+
+        if (isset($map['FileId'])) {
+            $model->fileId = $map['FileId'];
+        }
+
+        if (isset($map['Success'])) {
+            $model->success = $map['Success'];
+        }
+
+        return $model;
+    }
+}